+++ contracts/empty/src/lib.rs
@@ -1,0 +1,2 @@
+// Make sure to reexport cosmwasm_std, so the standard entry points are kept in the final wasm
+pub use cosmwasm_std;