+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to
   `Addr` type for `ContractInfoResponse::{creator, admin}`. ([#1883])
 - cosmwasm-std: Change `DistributionQuerier::new` to take `IntoIterator` instead
   of `HashMap`. ([#1941])
+- cosmwasm-vm: Make `instantiate` entrypoint optional. ([#1933])
 
 [#1874]: https://github.com/CosmWasm/cosmwasm/pull/1874
 [#1876]: https://github.com/CosmWasm/cosmwasm/pull/1876
@@ -54,6 +55,7 @@ and this project adheres to
 [#1884]: https://github.com/CosmWasm/cosmwasm/pull/1884
 [#1898]: https://github.com/CosmWasm/cosmwasm/pull/1898
 [#1902]: https://github.com/CosmWasm/cosmwasm/pull/1902
+[#1933]: https://github.com/CosmWasm/cosmwasm/pull/1933
 [#1939]: https://github.com/CosmWasm/cosmwasm/pull/1939
 [#1940]: https://github.com/CosmWasm/cosmwasm/pull/1940
 [#1941]: https://github.com/CosmWasm/cosmwasm/pull/1941